Import rating times correctly.  Export all to one csv file with binary genre variables
git-svn-id: svn://had.co.nz/projects/projects/movies@153 688fc0e7-72e3-0310-a89e-c7106b9812cc

+++ movies.sql
@@ -4,7 +4,7 @@ title varchar(250),
 year integer,
 budget integer,
 length integer,
-imdb_rating integer,
+imdb_rating float,
 imdb_votes integer,
 imdb_rating_votes varchar(10),
 mpaa_rating varchar(5)
